requirements.txt
absl-py==0.15.0
argon2-cffi==21.1.0
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
boto3==1.21.9
botocore==1.24.9
cachetools==4.2.4
captum==0.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
entropy==0.1.5
entrypoints==0.3
filelock==3.4.2
gdown==4.2.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.2
google-auth-oauthlib==0.4.6
grad-cam==1.3.7
grpcio==1.41.1
huggingface-hub==0.4.0
idna==3.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
oauthlib==3.1.1
opencv-python==4.5.4.58
packaging==21.2
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.21
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
pytorch-transformers==1.2.0
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.2
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.10
setproctitle==1.2.3
Shapely==1.8.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
timm==0.4.12
tokenizers==0.11.6
toml==0.10.2
tomli==1.2.3
torch==1.9.0
torchattacks==3.2.4
torchaudio==0.9.0
torchcam==0.3.1
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
ttach==0.0.3
typing-extensions==3.10.0.2
uncertainty-toolbox @ git+https://github.com/uncertainty-toolbox/uncertainty-toolbox@946433b2bca9eb93b06b144cffdb32faf0a9c64f
urllib3==1.26.7
wandb==0.12.15
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
